Update patch set 7
Patch Set 7:

Hi Peter. Thanks for the comments. As you say, there is probably more that can be done on scalability, but the purpose of this patch is just to fix the signalling issue we saw with the position the film grain is signalled for hidden frames. So I'm hoping we can get this one in and approved at the WG. Scalability will have to be taken up separately I think (this is not something I am planning to work on).

Patch-set: 7
Reviewer: Gerrit User 7984 <7984@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>