[CFL] Constant Propagated TX_SIZE for subsampling
Macros to generate subsampling functions for each
TX_SIZE supported by CfL. This allows the compiler
to inline a generic version of the function into
each variation. The benefit of this is that the
width and the height are constant inside this
function. The subsampling type is also constant
propagated, but only 4:2:0 is currently
implemented.

Change-Id: I16cb33d1a7bfd1ec57cd9f68cb17116068a8e76a