[NORMATIVE] remove reset_decoder_state signaling
FWD_KF previously explicitly signal the one-bit symbol of
reset_decoder_state. However, this tool has been redefined, that only
frames of KEY_FRAME types, not just INTRA_ONLY, can be coded as no-show
frame and buffered for the use of forward key framnes. Hence, when
show_existing_frame == 1, if the frame-to-show frame points to a
buffered identified as KEY_FRAME, the a forward key frame is used, and
the status of reset_decoder_state can be auto-filled to one.

Change-Id: Ife1c2a7f24b19073d6b809d524debae7fa4cacdd