[NORMATIVE] Add sequence header control for dual filter
Add a sequence header control to turn on and off the use of
dual direction interpolation filter. This allows the encoder to
select and code only one filter type per coding block, when
turning off the dual filter option.

Thanks to Ankur and the Nvidia team for identifying the potential
issues for hw encoder and proposing the solution.

BUG=aomedia:1326

Change-Id: Icfb82940d85c5f77406f613be7fa829763289514