Add encoder support on FWD_KF update
The status for reset_decoder_state is not needed to be explicitly
signaled, as the forward coded keyframe now has to be specified as
KEY_FRAME for its frame type. Whenever a show_existing_frame points
to a to-show frame as KEY_FRAME, it will be identified as forwarded
coded keyframe. Intra only frames are no longer considered as forward
coded keyframe.

Change-Id: I1b612dfe50e669f433aba4a3b5a47e378e4311ee