Add encoder support on FWD_KF update
The status for reset_decoder_state is not needed to be explicitly
signaled, as the forward coded keyframe needs to have KEY_FRAME
specifically as its frame type. Whenever a show_existing_frame points
to a to-show frame with KEY_FRAME as frame_type, it will be identified
as forwarded coded keyframe. Intra only frame cannot be regarded as
forward coded keyframe any more.

Change-Id: I1b612dfe50e669f433aba4a3b5a47e378e4311ee