[NORMATIVE-SYNTAX, obu-sizing] Fix length field size handling
It is possible, and reasonable from an encoder perspective, to
encode an OBU size field using more than the minimal number of
bytes. With the current reference decoder, such lengths are
decoded correctly, but the data pointer is then advanced by the
wrong number of bytes, leading to a desync.

This can easily be fixed by having aom_uleb_decode() return the
exact number of bytes read. The behaviour of returning -1 on error
can be preserved.

In addition, fix aom_uleb_encode_fixed_size(), and strengthen its
tests - this function appears to have been padding its output
incorrectly, but this was not caught by the test suite.

The 'encode' tests also mixed big-endian and little-endian in a
potentially confusing way (reading back the encoded bytes big-endian
for comparison to a 32-bit hex constant). Change to using an array
of expected bytes, to make things clearer.

BUG=aomedia:1433

Change-Id: Iea80a1b70299e0a26cd45e44ecb039804f8c5245