Add option to encoder to force max frame w/h
Prior to this patch: when writing the sequence header at the start of
each sequence, max_frame_width and max_frame_height were set from the
dimensions of the current frame. This posed a problem for a resizing
video source, where the dimensions of the frames could have increased
within a sequence.

This patch adds an option to the encoder to force the values of
max_frame_width and max_frame_height. This feature is then used to fix
the resize tests that otherwise caused errors.

BUG=aomedia:1431

Change-Id: Ic1d347a33ae8bf93c0ca5f631cfb1eeb09d25d30