Update patch set 2
Patch Set 2:

> > i) The fixed-size encode functions seemed incorrect - see my comment on the bug report for more
> 
> It's certainly possible-- they worked when in use, but that doesn't mean there were no bugs. The issue aomedia:1125 tracked the progression of the leb128 support. Note that it was never intended for use on a big endian system, so not surprised that appears to be an issue.

This isn't to do with the endian-ness of the system (in fact, all the code seems to correctly be independent of system endianness) - the previous code was reading 4 bytes in big-endian order so that, say, expected_bytes = {0x12, 0x34, 0x56, 0x78} corresponds to expected_value = 0x12345678

This is fine as far as it goes, it's just potentially confusing because the actual encode/decode process is always done in little-endian order.


From a quick look, it seems the reason it used to work (at least as of https://aomedia-review.googlesource.com/c/aom/+/42563 ) is that the old code was unconditionally skipping 4 bytes, regardless of how long the LEB128 decoder thought the field was.


> > ii) I'm not sure that obudec.c reads the OBU sizes correctly with obu-sizing enabled (independently of this patch), but I'm having difficulty working out exactly what the code is trying to do. Please could someone else have a look into this?
> 
> I'm working on obudec-- it's definitely broken w/CONFIG_OBU_SIZING turned on. It blindly uses PRE_OBU_SIZE_BYTES and pays no attention to the variable length size fields. WIP patch here: https://aomedia-review.googlesource.com/c/aom/+/47881

Okay - thanks for working on that! :)

Patch-set: 2
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>