Reset loop-filter parameters when intrabc is allowed
Explicitly reset the loop-filter parameters when intrabc is allowed for
current frame.

Before this patch, when encoding consecutive keyframes, there may be
mismatches. E.g., allow_intrabc is false for the first frame, and a
none-zero value is assigned to cdef_bits. Then allow_intrabc is true for
the second frame. On the encoder side, the none-zero value is carried
over to the second frame; but on the decoder side, it's set as 0.

To make the logic more clear and play safe, we reset all loop-filter
related parameters, not just cdef_bits.

BUG=aomedia:1429

Change-Id: Ic674b92c72450d5cf298303e17a3be4208c3cf85