Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> Instead of forcing max width/height explicitly, shouldn't the encoder figure it out internally?

Hi Urvang, this was my thinking:

What’s going on is when writing the sequence header the encoder is setting max_frame_width (resp. height) to the width (resp. height) of the current frame, i.e. the first frame in the sequence. This is a problem if the frames in the sequence get bigger. There are a few ways I can think of to solve this:

1) On the first pass, when determining the keyframes, we could also determine the maximum frame width and height in each sequence. Con: this won’t work when encoding is done in a single pass e.g. video conferencing, which is perhaps where this resizing video source is most useful.

2) Start a new sequence every time the frame increases in size. Con: this could result in an overall large increase in the encoded size if the frame size is fluctuating a lot (which perhaps also is likely to happen with video conferencing).

3) Give the option of forcing a maximum frame width and height in the encoder. Con: extra options for the encoder.

To me, number 3 seems to be the most sensible solution. What do you think?

Imdad

Patch-set: 1
Reviewer: Gerrit User 8213 <8213@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>