Remove multiple options for LOWPRECISION_BLEND
Previsouly CONFIG_LOWPRECISION_BLEND could be set as 1 or 2
to experiment with two different options, but 2 was what was
adopted. This patch removes the old 1 option and makes
CONFIG_LOWPRECISION_BLEND=1 work the way the adopted tool
is supposed to behave for consistency with the other tools and
avoid confusion.

To continue experiments with other options one will need to manually
change the macros ROUND0_BITS or COMPOUND_ROUND1_BITS in
common/convolve.h.

Change-Id: I6e5434bbce1e655380a35cb1b530c912c5145560