[NORMATIVE] Fix interaction of get_vartx_max_txsize() + txk-sel
The chroma transform size clamping applied by get_vartx_max_txsize()
was more restrictive than it needed to be, and this was causing a bad
interaction with the txk-sel experiment (see linked bug report).

For example, when using 4:2:2 subsampling, it is entirely valid for
a 64x64 luma transform to be subsampled to 32x64, but the code was
only allowing a 32x32 transform instead.

This patch rewrites get_vartx_max_txsize() so that it always returns
the largest valid transform size given the input block size and
subsampling method. This fixes the interaction with txk-sel, and should
provide a slight coding improvement due to enabling 32x64 and 64x32
chroma transforms.

BUG=aomedia:1450

Change-Id: Iea5ecc81b858d91d8028965ebf3b4db25a0dd0f6