[NORMATIVE] Sub8x8 inter pred buffer offset in scaled ref frame
Properly account for the buffer offset in the sub8x8 chroma inter
prediction on the scaled reference frame. The sub8x8 chroma block
is processed at the bottom right of the collocated luma block.
Hence it needs to offset the reference buffer pointer to map back
to the 8x8 luma region.

BUG=aomedia:978

Change-Id: I046e454fd8c2fc6113804d159ce09691c209e0cf