[NORMATIVE] Extend full-pixel motion vectors to 12 bits
This patch followed the previous patch "Add one more MV_CLASS for
integer MV" and implemented the extending of full-pixel MVs from
11 bits to 12 bits.

The total bits for MV is 15 bits, including 12 bits for full pixel
part and 3 bits for subpixel part. Also, fixed the mv_q4 used in
predictions so that no overflow occurs.

BUG=aomedia:966

Change-Id: I53b55146d5eba8f3363a8caadcc604c90baa01a0