Revert "[wedge/compound-segment, normative] Remove more rounding"
This reverts commit 7dbb005144d51c6aa13cd48726091adf63c5326e.

Reason for revert: This inadvertently changed the multipliers for
blending to be 2 bits more - which was not what was agreed in the WG.
Use of A256 vs A64 as in the reverted patch is non-normative for
luma (i.e. does not change luma), but for chroma it does. So it is
better to revert it to keep multiplers smaller.

Change-Id: Id1ddf44bf522dd390fc05c091c5c1b11eab9eff7