Update patch set 1
Patch Set 1: Code-Review-1

Modern compilers will generally ignore the INLINE keyword and decide for themselves whether or not to inline a function (especially if optimisations are turned on).

I compiled this patch and the preceding patch with both CLANG 4.0.1 and GCC 4.8.5 with the C flag O2 to find:
- GCC decided not to inline the functions in either case.
- CLANG inlined the functions in both cases.

Either way, both CLANG and GCC produced identical assembly code for this patch and for the preceding one. Therefore, I think that this patch probably adds unnecessary noise to the codebase.

If you really think that this patch is necessary, perhaps could you let me know the compiler version and flags that you used to compile the code?

Patch-set: 1
Reviewer: Gerrit User 8213 <8213@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=-1