Update patch set 3
Patch Set 3:

(1 comment)

The bug report is tagged as a specification bug, not as a feature enhancement, so I suspect most people will ignore it thinking it is just a mistake in the specification, when actually it is a request for a normative change.

By the way, has this change been discussed with the working group?  In the past, people have been unhappy about changes appearing in the code outside of an experiment flag.

Patch-set: 3