Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> There are a number of new test failures bisecting to this patch.

The issue seems to be from this patch, after the cmake doing a make clean before make, will cause a build failure due to removal of the av1_rtcd.h file. Before this patch it would still build. We need to update the bisect script to remove the make clean.

Patch-set: 8