[NORMATIVE, fwd-kf] Reload adapted CDFs for fwd keyframes
Instead of calling av1_setup_past_independence() when we display
a forward-coded keyframe, we can instead reload the frame context
which we had just after decoding that keyframe.

This is a much less fragile approach (it removes several issues
we encountered on our internal testing), and should give better
coding gains.

BUG=aomedia:1510

Change-Id: Ia67b20bc4128b85b11829bc84e0c383f36ae4a52