[NORMATIVE-SYNTAX] Fix max tile width/area calculation
As per the linked bug report, the maximum #pixels in each tile
was previously being computed in a way which depended on the
superblock size. This patch fixes this by making max_tile_width_sb
and max_tile_area_sb depend on the superblock size.

One potential source of confusion comes from the fact that we
already have a variable called max_tile_width_sb elsewhere in
the code, which tracks the width of the widest tile in the current
frame. To avoid confusion, this is renamed to widest_tile_sb,
so that max_* always means a limit in this context.

BUG=aomedia:1508

Change-Id: I94ebbc46e54cb5b1e8eda5f8ab3f0a3cd61698e8