Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 3:
> 
> (1 comment)
> 
> The bug report is tagged as a specification bug, not as a feature enhancement, so I suspect most people will ignore it thinking it is just a mistake in the specification, when actually it is a request for a normative change.
> 

I'm sorry about that. Can i just change the bug type now?

> By the way, has this change been discussed with the working group?  In the past, people have been unhappy about changes appearing in the code outside of an experiment flag.

This change wasn't formally discussed in the WG on it's own. But when it was presented in the context of another OBU related change, one of the comments were to do this as a normative change with a bug associated with it. This also didn't affect anything in a negative way and hence i thought an experiment flag wasn't necessary.

Patch-set: 3