Add droppable frame support in CONFIG_DROPPABLE_FRAMES
Adds an option to change the meaning of the error_resilient_mode
flag to mean droppable frames. Droppable frames are those
that can be dropped and the subsequent frames can still be
decoded without any loss.

This will be more useful than the previous error-resilient-mode
interpretation - which only meant subsequent frames could be
entropy-decoded but drift was still possible.

WIP.

Change-Id: I75798a55d6535f17ffb97519bde3ef3eb8298bc5