[NORMATIVE, fwd-kf] Signal refresh_frame_flags for fwd keyframes
Synchronize the fwd-kf experiment with its design document,
by explicitly signalling refresh_frame_flags for non-shown keyframes.

Note: The behaviour of setting refresh_frame_flags = 0xFF when we
show a previously-encoded keyframe is kept. We only change which
reference slot a non-shown keyframe overwrites *at the time when it
is decoded*

BUG=aomedia:1509

Change-Id: I3a4e90a0b9dceab25f94312bd4ecc96ea016f624