[NORMATIVE] Fix coefficient CDFs when using PRIMARY_REF_NONE
Fix a subtle decode order dependency introduced by the recent
changes to the no-frame-context-signaling experiment; see the
attached bug report for details.

In addition, by moving the call to av1_setup_past_independence()
to after we read cm->base_qindex, we avoid needing to re-setup
the coefficient CDFs and re-store the default frame context,
as we were doing before.

BUG=aomedia:1521

Change-Id: Id6319fe5cf9379f81876404cd52899fa2e03f519