Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Just curious why "cpi->common.bit_depth > 8" doesn't work?

Because the whole pipeline can be forced to work with 16-bit pixels, regardless of the bitdepth. This is what CONFIG_LOWBITDEPTH=0 does (The "16-bit" pipeline is actually the reference for decoding, while the 8-bit and SSE are considered as optimizations).

Patch-set: 1
Reviewer: Gerrit User 6000 <6000@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>