[normative] Robust order hint coding
Frame offsets are coded relative to the frame offset of a reference
picture (LAST) such that the frame offset prediction chain is
consistent with the picture prediction chain.
Frame offsets used for prediction are not updated for frames that
are not shown such as to guarantee that the frame offset delta
is always positive.

This is a proposed solution to the following bug
BUG=aomedia:1497

Change-Id: I76a5fb506344e8298bcc054f2013d2b717d6abce