[Normative] Add a bit of order hint in sequence header
Add a bit in sequence header to enable/disable use of frame order
hint. If this bit is 0, it will disable three tools that rely on
frame order hint:
(1) ref_frame_mvs, (2) JNT_COMP, (3) REF_FRAME_SIGN_BIAS.
By default it is set as 1.

The command line configure of this flag is not implemented yet.
If one needs to change it to zero for test, search default_extra_cfg
in av1_cx_iface.c.

Change-Id: Icf4f8643e37fef2486ef79c80787979c48208493