Update patch set 3
Patch Set 3:

Would this scheme work for hierarchical coding if we only want the overhead of error resilience for some frames?

For example, suppose we have frames at times 0,1,2,3 and have decided to encode the frames in this coding order:
Frame 0 as Key frame (in error resilience mode)
Frame 3 as P frame depending on frame 0 (in error resilience mode)
Frame 2 as P frame depending on frame 3 (in error resilience mode)
Frame 1 as B frame depending on frames 0 and 2

If frame 3 was dropped, my understanding is that 
Frame 0 would have correct OrderHint of 0
Frame 3 is dropped so OrderHint unknown
Frame 2 has OrderHint based on frame 3 so uses some random value

Then when we are trying to decode frame 1 we would have inconsistent OrderHints for its two reference frames, despite having received both of them in error resilient mode.

Is this a problem?

Patch-set: 3