Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Would this scheme work for hierarchical coding if we only want the overhead of error resilience for some frames?
> 
> For example, suppose we have frames at times 0,1,2,3 and have decided to encode the frames in this coding order:
> Frame 0 as Key frame (in error resilience mode)
> Frame 3 as P frame depending on frame 0 (in error resilience mode)
> Frame 2 as P frame depending on frame 3 (in error resilience mode)
> Frame 1 as B frame depending on frames 0 and 2
> 
> If frame 3 was dropped, my understanding is that 
> Frame 0 would have correct OrderHint of 0
> Frame 3 is dropped so OrderHint unknown
> Frame 2 has OrderHint based on frame 3 so uses some random value
> 
> Then when we are trying to decode frame 1 we would have inconsistent OrderHints for its two reference frames, despite having received both of them in error resilient mode.
> 
> Is this a problem?

I guess that in such a case frame 2 would have to be coded as dependent on frames 3 and 0, and derive its OrderHint from frame 0.
Assuming there is no issue with OrderHint would it actually be possible to correctly parse frame 1 if the proper scaling of MVs in frame 2 is unknown (because frame 3 was dropped)?

Patch-set: 3
Reviewer: Gerrit User 8487 <8487@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>