Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > Because frame 2 is sent as error_resilient, I believe it should be possible (once the other error_resilience bugs I reported are fixed) to compute all the frame 2 motion vectors correctly.
> 
> Assuming MVs for frame 2 are indeed correctly reconstructed, wouldn't motion_field_projection() produce incorrect data when decoding frame 1?

That is a very good point and another error resilience problem that needs to be fixed.
I think we need to ensure that error_resilience frames have all state correct (except for pixel data).  The extra bit of state here is called SavedOrderHints in the specification and I think it needs to be cleared for error_resilient frames.

Patch-set: 3