Update patch set 3
Patch Set 3:

In order to correctly decode a frame that relies on OrderHint with this scheme, you need to be able to trace the OrderHint of all of its references back to a common ancestor: either a keyframe, or a single error-resilient frame. That's easy to do by just tagging each reference with the source of its OrderHint value.

For example, you can set the tag to 0 for key frames and then pick a unique ID (i.e., increment a counter) for each error-resilient frame you decode that is missing the reference it needs to derive OrderHint. Frames that have the reference they need to derive OrderHint copy the tag from that reference. Non-error-resilient frames that are missing a reference cannot be decoded (so the tagging doesn't matter, though you could in theory parse the frame header and propagate OrderHint values through such frames as well).

So, in Peter's example, Frame 0 is tagged with 0. Frame 2 is tagged with 1 (because it is missing Frame 3). Then at Frame 1, you detect that not all of your references have the same tag and deduce that you cannot parse Frame 1.

So it's possible to know if you can parse a frame with relatively little work. With Peter's proposal, you would actually be able to parse Frame 1, since OrderHint is not coded as a delta, so that's a potential benefit of that scheme.

Patch-set: 3
Reviewer: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>