[NORMATIVE] Signal ref OrderHints in err res mode
When a frame is marked as error_resilient, the OrderHints of all of its
reference frames are explicitly signalled in the bit stream.

On the decoder side, it is asserted that the values read from the bit
stream match the values in the reference buffers. In addition, it is
asserted that all reference frames point to valid buffers.

This is the first patch in a series of patches affecting
error_resilient_mode.

Change-Id: Icfbdd139dd4de01287d7e5b19309506305b854cc