Allocate missing buffers in error_resilient_mode
(When a frame is marked as error_resilient, the OrderHints of all of its
reference frames are explicitly signalled in the bit stream.)

On the decoder side:

- If a reference frame points to an invalid frame buffer, we allocate a
  new frame buffer with default pixel values and assign it to this
  reference frame instead. The OrderHint (aka cur_frame_offset) of this
  new buffer is set to the value read from the bit stream.

- Otherwise, we assert that the value read from the bit stream matches
  the OrderHint (aka cur_frame_offset) of the buffer.

This is the latter of two patches affecting error_resilient_mode. It is
a non-normative change that improves the ability of the decoder to cope
with invalid bitstreams.

Change-Id: Ic465feb90c6aeb07b732d0bf28ad14ddc5c304f7