Update patch set 1
Patch Set 1:

Although I am not in favour of this patch (largely because it makes my brain hurt to think through all the consequences), it does address an independent problem to OrderHints.

The issue is with the following sequence:
1) frame that is lost
2) error resilient frame
3) normal frame that depends on frames 1 and 2

The explicit order hints mean that we know the correct value of OrderHint for frame 1.
However, we still don't know its motion vectors.
This means that we cannot use MFMV to decode frame 3.

With this patch, it does become possible because MFMV is automatically disabled.

My preference (on grounds of simplicity) is that we get the same functionality by simply changing the encode to set syntax element can_use_previous equal to 0 to disable MFMV.

However, I do agree that this patch would have the potential for improved coding performance as it only partially disables MFMV.

Patch-set: 1
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>