Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I agree that the encoder can already do this. The fact that it makes your brain hurt to think through the consequences is a good reason not to do it. You end up having to add another signalling layer to make these dependencies explicit. This was the issue with MMCO with H264.

Thanks all for all the comments. Indeed 'brain hurting going through all the use cases and consequences is a nice way of putting it'.

We will certainly fix all the issues raised.
However, I think the key point is to actually decide what behavior is expected/wanted of the error-resilient-mode bit. Fixing bugs with the implementation comes next. 

Before this patch, this is the behavior we had:
Any frame marked as error-resilient shall be parse-able by itself. But subsequent frames not marked as error-resilient may still not be parse-able.

The intent of the patch was to make the behavior somewhat more powerful:
Any frame marked as error-resilient shall not only be parse-able by itself, but all subsequent frames even those not marked as error-resilient should be parse-able as well.

However, after some back and forth discussions we realize it may be best to have the latter behavior only for the base layer error-resilient frames. All enhancement layer error-resilient frames will go back to the original behavior. Do any of you see any issues with that?

We hope to have a session on Error-Resilience on next Friday's WG call. Hopefully all the code will be correctly implemented by then.

Patch-set: 1
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>