Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> As Thomas Davies pointed out, this patch breaks far more than it fixes.
> 
> With this change, if you lose *any frame at all* you never know if you can correctly parse (much less decode) any future frame (in decode order). Because that frame *might* have been error resilient, and made state changes you have no other way of finding out about.
> 
> The correct invariants to maintain are that if you have received and correctly parsed all of your reference frames, then you should be able to parse your current frame, and if you have correctly decoded all of your reference frames, then you should be able to correctly decode your current frame (even if it is not error resilient in any way).

There may be a misunderstanding here.
If we lose any frame (error-resilient or not) we cannot decode anything until the next error-resilient frame. That behavior is not changed by this patch. 
There are issues with this patch that are being fixed, but intent-wise the behavior is not changed.

If we want to achieve the functionality of a 'droppable' frame, that is achieved by setting refresh_frame_flags=0, which is different from error_resilient_mode.

Patch-set: 1
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>