Update patch set 1
Patch Set 1:

> There may be a misunderstanding here.

I agree that there is.

 > If we lose any frame (error-resilient or not) we cannot decode
 > anything until the next error-resilient frame. That behavior is not
 > changed by this patch.

So, this has not been the intent of the design for the past two years. That is the whole purpose of the experiments like NO_FRAME_CONTEXT_SIGNALING and the other changes to global MV signaling, etc. So, if I have correctly decoded all of my reference frames, what prevents me from being able to decode my current frame, even if I have not received other frames (that I am not referencing)? Because we need to fix that.

 > If we want to achieve the functionality of a 'droppable' frame,
 > that is achieved by setting refresh_frame_flags=0, which is
 > different from error_resilient_mode.

No, this is a special case. We want the decoder, looking at only the frames it has received, to *know* if it can successfully decode a frame. That way it can never present a broken frame to the user. Setting refresh_frame_flags=0 should certainly enable a frame to be droppable, but you may also wish to drop other frames (spatial or temporal scalability layers, or other more flexible structures based on real-time feedback), which were stored in the reference buffer, but which might not be referenced by some frames (which should still remain decodable).

Patch-set: 1
Reviewer: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>