Update patch set 2
Patch Set 2:

> What behavior do you want to see from the error-resilient-mode ? It
 > all boils down to that?

What I want is for error resilient frames not to break the invariants I described in the other patch. I want a decoder to be able to know what it is able to parse and what it is able to decode based solely on the things it has received.

If you want all frames to be parseable after an error resilient frame, then make the encoder only use that error resilient frame (and subsequent frames) as your references.

Patch-set: 2
Reviewer: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>