Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > Why is this needed? I think it is not needed to parse the current frame if it is error resilient. Is this intended to assist in resetting things for future frames?
> > 
> > What is the estimated overhead of adding this signalling, say at 60 Hz, 400kb/s?
> 
> At the moment there are several features (such as jnt_comp and short_signalling) that cannot be used in error_resilience frames because they depend on order hints for the reference frames.
> 
> The idea is to re-enable the use of these features by explicitly sending the order hints.
> 
> Hopefully the gains will be greater than the overhead.

But if a decoder has correctly received the reference frame then it has the order hint for that reference frame as it is being explicitly coded now. If it hasn't got the reference frame, it can't decode anyhow. 

Are there additional parsing dependencies that are dependent on having the order hint for a reference frame that has not been received?

Patch-set: 6