Update patch set 6
Patch Set 6:

> Are there additional parsing dependencies that are dependent on having the order hint for a reference frame that has not been received?

Yes, there are several dependencies I have reported under issue 1497 when we are attempting to decode an error resilience frame.
For example, whether compound prediction is enabled depends on order hints for the reference frames.

There is also a more subtle dependency when we have:
1) Error resilient frame
2) Non-error resilient frame

Due to the MFMV experiment, the parsing of frame 2 depends on the order hints of references of frame 1! (due to the motion vector scaling)

The stored state for each reference buffer includes both the OrderHint for that frame, and an array of the OrderHints for each referenced frame.

Patch-set: 6
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>