Fix syntax sync issue for show_existing_frame = 1
Prior to this patch, a zero byte was written to the bitstream in
write_uncompressed_header_obu() when cm->show_existing_frame == 1
and cm->seq_params.frame_id_numbers_present_flag == 1. However,
this byte is not read in read_uncompressed_header(), resulting in
a mismatch in header sizes.

This patch resolves this issue by reading and discarding the zero
byte in read_uncompressed_header().

BUG=aomedia:1579

Change-Id: I78f722aa46ca4a5c3a6a571e9e17baa0ea1a1dcb