[NORMATIVE] Take segment features from LAST_FRAME
This fixes bug #1441, which described a situation where not updating the
segment data meant reusing the previously decoded frame's segment data whilst
the segmentation map used was taken from LAST_FRAME (as long as the
CONFIG_SEGMENT_PRED_LAST experiment is turned on, which it is by default).

This commit extends the CONFIG_SEGMENT_PRED_LAST experiment by storing the
segmentation features alongside the segmentation map and reusing them by
copying them to the main segmentation struct in AV1_COMMON if the update_data
flag is turned off but segmentation is enabled.

Change-Id: Iec094e6989aefc8587541dd7e9230ae69c735ecb