Update patch set 1
Patch Set 1: Code-Review-1

No, its the responsibility of the encoder to ensure that if segmentation is turned on no qindex will go to 0 unless lossless mode is switched on.
We should not be fixing this with such hacks but by making the encoder smarter.

Patch-set: 1
Reviewer: Gerrit User 6510 <6510@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=-1