Add film grain table input to allow injecting grain params.
In typical scenario where film grain renoising is done, the denoiser
will be applied to the input and the denoised source will be encoded.
This film grain table is meant to map a timestamp to corresponding
noise parameters, and can be provided along with the denoised source.

In an online denoising (within the encoder) the table can be used
to enqueue film grain parameters that can then be dereference when
the encoded frames are written out.

Change-Id: Iaec82a6df2c73da31d75c04d08fdf210607f1ba8