[NORMATIVE] Always update segfeatures at intra-only frames
When we are decoding an intra-only frame, cm->prev_frame is not
meaningful (it points to the previous decoded frame, which we
don't want to depend on). So we shouldn't be loading data
from cm->prev_frame->seg. Similarly, for an error-resilient
frame, the frame corresponding to cm->prev_frame may have
never been received, so we can't depend on it for anything
affecting the frame syntax.

Instead, in both of these cases, we should always send a fresh
segment feature list, ie. force seg->update_data = 1.

Also refactor things slightly, so that the separation between the
intra-only/error-resilient path and the regular path is clearer.

BUG=aomedia:1441

Change-Id: Ifa7bb08ab542406ce4be64887741f60358f3e6ef