Clear frame distance interpretation in explicit_order_hint
When the explicit_order_hint is turned on, the frame index will
take cyclic count between 0 - 127. The frame distance computation
will interpret distance (80 - 0) as (80 - 128) -> -48. For many
1-pass video conferencing settings, one might for example keep
a golden or arf frame for a long time, which would result a frame
distance computation from 80 to 0 as -48 and causing codec failure
in mfmv projection.

To solve such potential mis-interpretation, we always drop the
negative frame distance in the denominator part of the motion vector
projection.

BUG=aomedia:1593
BUG=aomedia:1599

Change-Id: I6a6fe32c2d9398c12195cf1648ccabdc56dab6d5